+++ front/components/assistant_builder/shared.ts
@@ -1,16 +1,12 @@
 import type {
-  CoreAPITable,
   DataSourceType,
-  DataSourceViewType,
   LightContentNode,
   TimeframeUnit,
 } from "@dust-tt/types";
 import type { ConnectorProvider } from "@dust-tt/types";
 import {
-  getGoogleSheetContentNodeInternalIdFromTableId,
   getGoogleSheetTableIdFromContentNodeInternalId,
   getMicrosoftSheetContentNodeInternalIdFromTableId,
-  getNotionDatabaseContentNodeInternalIdFromTableId,
   getNotionDatabaseTableIdFromContentNodeInternalId,
   isGoogleSheetContentNodeInternalId,
 } from "@dust-tt/types";
@@ -310,34 +306,6 @@ export function getTableIdForContentNode(
     // For static tables, the tableId is the contentNode internalId.
     case null:
       return contentNode.internalId;
-
-    default:
-      throw new Error(
-        `Provider ${dataSource.connectorProvider} is not supported`
-      );
-  }
-}
-
-export function getContentNodeInternalIdFromTableId(
-  dataSourceView: DataSourceViewType,
-  table: CoreAPITable
-): string {
-  const { dataSource } = dataSourceView;
-  const { table_id: tableId } = table;
-
-  switch (dataSource.connectorProvider) {
-    case "google_drive":
-      return getGoogleSheetContentNodeInternalIdFromTableId(tableId);
-
-    case "notion":
-      return getNotionDatabaseContentNodeInternalIdFromTableId(tableId);
-
-    case "microsoft":
-      return getMicrosoftSheetContentNodeInternalIdFromTableId(tableId);
-
-    // For static tables, the contentNode internalId is the tableId.
-    case null:
-      return tableId;
 
     default:
       throw new Error(
