+++ front/lib/api/data_source_view.ts
@@ -9,6 +9,7 @@ import { ConnectorsAPI, CoreAPI, Err, Ok, removeNulls } from "@dust-tt/types";
 import assert from "assert";
 
 import config from "@app/lib/api/config";
+import { getContentNodeInternalIdFromTableId } from "@app/lib/api/content_nodes";
 import type { OffsetPaginationParams } from "@app/lib/api/pagination";
 import type { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import logger from "@app/logger/logger";
@@ -212,10 +213,10 @@ export async function getContentNodesForStaticDataSourceView(
       tablesRes.value.tables.map((table) => ({
         dustDocumentId: table.table_id,
         expandable: false,
-        internalId: table.table_id,
+        internalId: getContentNodeInternalIdFromTableId(dataSourceView, table),
         lastUpdatedAt: table.timestamp,
         parentInternalId: null,
-        parentInternalIds: [],
+        parentInternalIds: table.parents,
         permission: "read",
         preventSelection: false,
         sourceUrl: null,
